Fix NullPointerException when obtaining the number of matched strings
Change Generex.matchedStringsSize() to call Generex.buildRootNode() so
that the rootNode, used to obtain the number of matched strings, is
initialised. Document that the method now throws StackOverflowError, as
the initialisation of the rootNode might fail with that exception.
Add test to validate that the method returns the expected numbers, for
the regexes already being tested.
Remove a statement no longer needed from a test method, called to
prevent the NullPointerException, which initialised the rootNode as side
effect.