Allow to query if (the pattern/Automaton of) Generex is infinite
Add instance method "boolean Generex.isInfinite()" that allows to query
if the pattern/Automaton of Generex is (or not) infinite. The method
uses the state of Automaton to know if the pattern is infinite.
Add class GenerexUnitTest to test the behaviour of the new method.
Add test dependency org.hamcrest:hamcrest-junit:2.0.0.0 (used by the new
tests).
The change allows to check, at runtime, if a given pattern is infinite,
for example, to not attempt to get all generated Strings, or, disallow
the use of infinite patterns.