Fix GenerexIterator implementation to always return the expected strings
The current implementation of the iterator does not return all possible
strings (but repeated strings instead). For example, with the regular
expression "a{3,6}" it returns "aaa", "aaaaa", "aaaaaa" and "aaaaaa",
missing the string with 4 letters (and returning with 6, twice).
Re-implement GenerexIterator basing on the Automaton core algorithm
SpecialOperations.getFiniteStrings(Automaton,int) but, non-recursive, to
avoid potential StackOverflowErrors, and supporting infinite patterns.
Add tests to assert the expected behaviour.
Uncomment an existing test, in GenerexIteratorTest, that was testing
that the iterator returns the strings in same order (and add an assert
for the expected number) as Generex.getMatchedString(int), which now it
does.
Remove class TransitionLevel, no longer used by GenerexIterator.