Allow to validate if a pattern is valid for Generex
Add class method "boolean Generex.isValidPattern(String)" that allows to
validate if a regular expression is a valid pattern for Generex. The
method attempts to create a RegExp instance, if it does not succeed the
pattern is invalid, otherwise it's valid (and expected to work when
instantiating a Generex instance).
Extract a method in Generex that creates a RegExp instance, called by
the constructor Generex(String) and the validation method, the extracted
method is also responsible to replace the predefined character classes.
As part of the extraction of the method the instance variable
predefinedCharacterClasses was changed to a (immutable) class constant
(which prevents unnecessary instantiations per Generex instance).
Add tests to assert the expected behaviour.
The change allows to check, at runtime, if a given pattern is invalid,
for example, to inform and prevent the creation of a Generex instance
(which would fail).