Overload Generex ctor to accept Random
This is useful for applications that want to use Generex but control the
source of randomness. junit-quickcheck, in particular, could use this
so that a generator of strings that match a particular pattern can be made,
using the source of randomness that junit-quickcheck passes to its generators,
rather than having a separate source of randomness for Generex, which would
hinder repeatability of tests.